Rename file to isSignedIn to match function name
