+++ packages/mgt-element/src/utils/isSignedIn.ts
